requirements.txt
pandas
streamlit
json
google-generativeai